Move distribution.ts to /codeql-cli/distribution.ts
