+++ extensions/ql-vscode/src/extension.ts
@@ -54,7 +54,7 @@ import {
   FindDistributionResultKind,
   GithubApiError,
   GithubRateLimitedError,
-} from "./distribution";
+} from "./codeql-cli/distribution";
 import {
   showAndLogErrorMessage,
   showAndLogExceptionWithTelemetry,
